+++ src/main/java/com/example/study_system/controller/QuestionInfoController.java
@@ -32,10 +32,14 @@ public class QuestionInfoController extends BaseController {
 	@RequestMapping(value = "/question", method = RequestMethod.POST)
 	public ResultDTO addQuestion(@RequestParam("question") String question,
 			@RequestParam("questionOption") String questionOptions) {
-		List<JQuestionOption> options = JSON.parseArray(questionOptions, JQuestionOption.class);
-		QuestionInfoWithBLOBs questionInfo = JSON.parseObject(question, QuestionInfoWithBLOBs.class);
-		int result = serviceFacade.getQuestionInfoService().addQuestion(questionInfo, options);
-		return success(result);
+		if (question == null || questionOptions == null) {
+			return validationError();
+		} else {
+			List<JQuestionOption> options = JSON.parseArray(questionOptions, JQuestionOption.class);
+			QuestionInfoWithBLOBs questionInfo = JSON.parseObject(question, QuestionInfoWithBLOBs.class);
+			int result = serviceFacade.getQuestionInfoService().addQuestion(questionInfo, options);
+			return success(result);
+		}
 	}
 
 	/**
@@ -47,7 +51,11 @@ public class QuestionInfoController extends BaseController {
 
 	@RequestMapping(value = "/{questionId}", method = RequestMethod.DELETE)
 	public ResultDTO deleteQuestion(@PathVariable("questionId") Long questionId) {
-		return success(serviceFacade.getQuestionInfoService().deleteQuestion(questionId));
+		if (questionId == null) {
+			return validationError();
+		} else {
+			return success(serviceFacade.getQuestionInfoService().deleteQuestion(questionId));
+		}
 	}
 
 	/**
@@ -59,12 +67,15 @@ public class QuestionInfoController extends BaseController {
 	 */
 	@RequestMapping(value = "/{questionId}", method = RequestMethod.POST)
 	public ResultDTO updateQeustionInfo(@PathVariable("questionId") Long questionId,
-			@RequestParam("question") String question, @RequestParam("questionOption") String questionOptions,
-			@RequestParam("count") Integer count) {
-		List<JQuestionOption> options = JSON.parseArray(questionOptions, JQuestionOption.class);
-		QuestionInfoWithBLOBs questionInfo = JSON.parseObject(question, QuestionInfoWithBLOBs.class);
-		int result = serviceFacade.getQuestionInfoService().updateQuestion(questionInfo, options, count);
-		return success(result);
+			@RequestParam("question") String question, @RequestParam("questionOption") String questionOptions) {
+		if (questionId == null || question == null || questionOptions == null) {
+			return validationError();
+		} else {
+			List<JQuestionOption> options = JSON.parseArray(questionOptions, JQuestionOption.class);
+			QuestionInfoWithBLOBs questionInfo = JSON.parseObject(question, QuestionInfoWithBLOBs.class);
+			int result = serviceFacade.getQuestionInfoService().updateQuestion(questionInfo, options);
+			return success(result);
+		}
 	}
 
 	/**
@@ -77,12 +88,17 @@ public class QuestionInfoController extends BaseController {
 	@RequestMapping(value = "/questions", method = RequestMethod.GET)
 	public ResultDTO selectQuestion(@RequestParam(value = "pageNum") Integer pageNum,
 			@RequestParam(value = "pageSize") Integer pageSize,
-			@RequestParam(value = "content", required = false) String content) {
-
-		PageInfo<QuestionResultDTO> result = serviceFacade.getQuestionInfoService().selectQuestion(pageNum, pageSize,
-				content);
-
-		return success(result);
+			@RequestParam(value = "content", required = false) String content,
+			@RequestParam(value = "questionType", required = false) Integer questionType) {
+//			@RequestParam("questionType") Integer questionType) {
+//		if (pageSize == null || content == null || pageNum == null || questionType == null) {
+		if (pageSize == null || content == null || pageNum == null) {
+			return validationError();
+		} else {
+			PageInfo<QuestionResultDTO> result = serviceFacade.getQuestionInfoService().selectQuestion(pageNum,
+					pageSize, content, questionType);
+			return success(result);
+		}
 	}
 
 	/**
@@ -93,7 +109,11 @@ public class QuestionInfoController extends BaseController {
 	 */
 	@RequestMapping(value = "/{questionId}", method = RequestMethod.GET)
 	public ResultDTO selectQuestionDetailed(@PathVariable("questionId") Long questionId) {
-		return success(serviceFacade.getQuestionInfoService().selectQuestionTitle(questionId));
+		if (questionId == null) {
+			return validationError();
+		} else {
+			return success(serviceFacade.getQuestionInfoService().selectQuestionTitle(questionId));
+		}
 	}
 
 	/**
@@ -104,7 +124,11 @@ public class QuestionInfoController extends BaseController {
 	 */
 	@RequestMapping(value = "/count", method = RequestMethod.GET)
 	public ResultDTO selectQuestionCount(@RequestParam(value = "questionType", required = false) Integer questionType) {
-		return success(serviceFacade.getQuestionInfoService().selectQuestionCount(questionType));
+		if (questionType == null) {
+			return validationError();
+		} else {
+			return success(serviceFacade.getQuestionInfoService().selectQuestionCount(questionType));
+		}
 	}
 
 	/**
@@ -115,7 +139,11 @@ public class QuestionInfoController extends BaseController {
 	 */
 	@RequestMapping(value = "/answer", method = RequestMethod.GET)
 	public ResultDTO selectAnalysis(@RequestParam("questionId") Long questionId) {
-		return success(serviceFacade.getQuestionInfoService().selectAnalysis(questionId));
+		if (questionId == null) {
+			return validationError();
+		} else {
+			return success(serviceFacade.getQuestionInfoService().selectAnalysis(questionId));
+		}
 	}
 
 }
